requirements.txt
fastapi
uvicorn
streamlit
google-genai
requests
pydantic
python-dotenv